requirements.txt
ctransformers==0.2.5
sentence-transformers==2.2.2
pinecone-client
langchain
langchain-pinecone
langchain-community
flask
pypdf
python-dotenv
-e .